DEX-934 refactor usePostSettingsAsset
